+++ connector/src/test/java/it/eng/connector/integration/s3/PartDownloaderTest.java
@@ -1,8 +1,10 @@
 package it.eng.connector.integration.s3;
 
+import it.eng.connector.integration.BaseIntegrationTest;
 import it.eng.datatransfer.model.DataAddress;
 import it.eng.datatransfer.model.TransferProcess;
 import it.eng.datatransfer.model.TransferState;
+import it.eng.datatransfer.repository.TransferProcessRepository;
 import it.eng.datatransfer.service.api.strategy.HttpPullTransferStrategy;
 import it.eng.datatransfer.util.DataTransferMockObjectUtil;
 import it.eng.tools.s3.configuration.S3ClientProvider;
@@ -13,10 +15,8 @@ import it.eng.tools.s3.service.BucketCredentialsService;
 import it.eng.tools.s3.service.S3BucketProvisionService;
 import it.eng.tools.s3.service.S3ClientService;
 import okhttp3.OkHttpClient;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.ContentDisposition;
 import org.springframework.http.MediaType;
 
@@ -25,9 +25,9 @@ import java.io.InputStream;
 import java.time.Duration;
 import java.util.UUID;
 
-@SpringBootTest
-@Disabled("Disabled for manual testing purposes. Enable only when needed.")
-public class PartDownloaderTest {
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+public class PartDownloaderTest extends BaseIntegrationTest {
 
     @Autowired
     TransferStateRepository stateRepository;
@@ -56,17 +56,14 @@ public class PartDownloaderTest {
     String sourceFileName = "s3.md";
     String destinationFileName = "bigFile.zip";
     String fileContent = "Hello, World!";
-
-    @Test
-    public void createBuckets() {
-        BucketCredentialsEntity sourceBucketCredentialsEntity = s3BucketProvisionService.createSecureBucket("dsp-true-connector-source");
-        BucketCredentialsEntity destinationBucketCredentialsEntity = s3BucketProvisionService.createSecureBucket("dsp-true-connector-destination");
-    }
+    @Autowired
+    private TransferProcessRepository transferProcessRepository;
 
     @Test
     public void testPartDownloader() throws InterruptedException {
-        BucketCredentialsEntity sourceBucketCredentialsEntity = bucketCredentialsService.getBucketCredentials("dsp-true-connector-source");
-        BucketCredentialsEntity destinationBucketCredentialsEntity = bucketCredentialsService.getBucketCredentials("dsp-true-connector-destination");
+
+        BucketCredentialsEntity sourceBucketCredentialsEntity = s3BucketProvisionService.createSecureBucket("dsp-true-connector-source");
+        BucketCredentialsEntity destinationBucketCredentialsEntity = s3BucketProvisionService.createSecureBucket("dsp-true-connector-destination");
 
         ContentDisposition contentDisposition = ContentDisposition.attachment()
                 .filename(sourceFileName)
@@ -81,17 +78,16 @@ public class PartDownloaderTest {
                             contentDisposition.toString())
                     .get();
         } catch (Exception e) {
-            System.out.println("Error uploading file: " + e.getMessage());
+            System.out.println("Error uploading file to source bucket: " + e.getMessage());
         }
 
         String presignURL = s3ClientService.generateGetPresignedUrl(sourceBucketCredentialsEntity.getBucketName(),
-                "minikube-installer.exe", Duration.ofDays(1));
+                objectKey,
+                Duration.ofDays(1));
 
-        System.out.println(presignURL);
         DataAddress dataAddress = DataAddress.Builder.newInstance()
                 .endpoint(presignURL)
                 .endpointType(DataTransferMockObjectUtil.ENDPOINT_TYPE)
-                .endpointType("")
                 .build();
 
         TransferProcess tp = TransferProcess.Builder.newInstance()
@@ -105,26 +101,11 @@ public class PartDownloaderTest {
         s3Properties.setBucketName("dsp-true-connector-destination");
         httpPullTransferStrategy.transfer(tp).join();
 
-//        S3ClientRequest s3ClientRequest = S3ClientRequest.from(s3Properties.getRegion(),
-//                null,
-//                destinationBucketCredentialsEntity);
-//        S3AsyncClient s3AsyncClient = s3ClientProvider.s3AsyncClient(s3ClientRequest);
-//
-//        PresignedBucketDownloader downloader = new PresignedBucketDownloader(stateRepository,
-//                s3AsyncClient,
-//                httpClient,
-//                "test-transfer-id-big_file.zip",
-//                presignURL,
-//                destinationBucketCredentialsEntity.getBucketName(),
-//                destinationFileName,
-//                destinationFileName);
-//
-//        System.out.println("Starting download...");
-//        downloader.run();
-    }
+        Thread.sleep(2000);
 
-    private String bucketEntityToString(BucketCredentialsEntity entity) {
-        return String.format("BucketCredentialsEntity{id='%s', bucketName='%s', accessKey='%s', secretKey='%s' }",
-                entity.getBucketName(), entity.getBucketName(), entity.getAccessKey(), entity.getSecretKey());
+        String presignURLDestination = s3ClientService.generateGetPresignedUrl(destinationBucketCredentialsEntity.getBucketName(),
+                tp.getId(),
+                Duration.ofDays(1));
+        assertNotNull(presignURLDestination);
     }
 }