+++ tools/src/main/java/it/eng/tools/s3/repository/TransferStateRepository.java
@@ -1,9 +1,9 @@
 package it.eng.tools.s3.repository;
 
-import it.eng.tools.s3.model.TransferState;
+import it.eng.tools.s3.model.TransferArtifactState;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TransferStateRepository extends MongoRepository<TransferState, String> {
+public interface TransferStateRepository extends MongoRepository<TransferArtifactState, String> {
 }